requirements.txt
selenium
pytest
allure-pytest
openpyxl
pandas
webdriver-manager
pytest-rerunfailures
pytest-xdist
